requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
pydantic==2.5.0
crawl4ai==0.5.0
pytest==7.4.0
beautifulsoup4==4.12.2
selenium==4.18.1
webdriver-manager==4.0.2
openai==1.73.0
numpy==1.26.4
google-generativeai==0.5.0
streamlit==1.36.0
flask==2.3.3
flask-wtf==1.2.1
flask-bootstrap5==0.1.1
jinja2==3.1.2
werkzeug==2.3.7
